% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{benchmark_worst_fdr}
\alias{benchmark_worst_fdr}
\title{Calculate FDR-corrected power for a range of phases and frequencies}
\usage{
benchmark_worst_fdr(
  mt,
  Amp,
  f0,
  Nmc,
  Nacro,
  Nfreq,
  fmin,
  fmax,
  fdr_method = "BH"
)
}
\arguments{
\item{mt}{vector of measurement times}

\item{Amp}{amplitude of signal}

\item{f0}{true frequency of signal}

\item{Nacro}{number of acrophases to use in simulating the power}

\item{Nfreq}{number of candidate frequencies to try in harmonic regression}

\item{fmin}{lowest candidate frequency}

\item{fmax}{highest candidate frequency}

\item{fdr_method}{passed to the \link{p.adjust} function, this determines what type
of multiple test correction should be performed}
}
\value{
the worst-case FDR-corrected power
}
\description{
Given a measurement schedule, true frequency, and amplitude, the FDR-corrected power
is estimated for each acrophase and the minimum power is returned.
}
\note{
Multiple test correction is performed only with respect to the multiple frequencies and
not with respect to the multiple phases.
}
\author{
Turner Silverthorne
}
