% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfun_sim_anneal_pwr.R
\name{cfun_sim_anneal_pwr}
\alias{cfun_sim_anneal_pwr}
\title{Cost function for simulated annealing power}
\usage{
cfun_sim_anneal_pwr(x, Amat, opts)
}
\description{
Cost function for simulated annealing power
}
\author{
Turner Silverthorne
}
