% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfun_2lattice_svdpower.R
\name{costfun_2lattice_svdpower}
\alias{costfun_2lattice_svdpower}
\title{Helper function for continuous power optimization, restricted to two lattice subspace}
\usage{
costfun_2lattice_svdpower(
  shift1,
  shift2,
  scale1,
  scale2,
  lat1,
  lat2,
  freqs,
  Amp = 1,
  alpha = 0.05
)
}
\description{
Helper function for continuous power optimization, restricted to two lattice subspace
}
