% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfuns.R
\name{costfun_2lattice_svdpower_discrete}
\alias{costfun_2lattice_svdpower_discrete}
\title{Cost function for 2-lattice discrete-time power optimisation}
\usage{
costfun_2lattice_svdpower_discrete(N1, dx1, N2, dx2, xshift2, tau, freqs, ...)
}
\arguments{
\item{N1}{number of points in first grid}

\item{dx1}{spacing of first grid}

\item{N2}{number of points in second grid}

\item{dx2}{spacing of second grid}

\item{xshift2}{phase shift of second grid}

\item{tau}{vector of candidate measurement times}

\item{freqs}{frequencies at which cost function should be evaluated}
}
\value{
minimum of cost function evaluated at \code{freqs}
}
\description{
Similar function to \link{costfun_svdpower} except measurement times can only
be collected at union of two uniform grids and these grids are selected from
a discrete underlying uniform grid.
}
\note{
This function works in units where \code{dx1} and \code{dx2} are assumed
to refer to indices in the \code{tau} grid.

Remaining arguments are passed to \link{costfun_svdpower}
}
\author{
Turner Silverthorne
}
