% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfuns.R
\name{costfun_auglattice}
\alias{costfun_auglattice}
\title{Cost function for 2-lattice continuous-time power optimisation}
\usage{
costfun_auglattice(N1, N2, shift2, scale2, freqs, ...)
}
\description{
Very similar to \link{costfun_2lattice_svdpower_discrete} except now, one of the lattices
is assumed to be fixed at the roots of unity.

This is useful because it removes degeneracy from the problem.
}
\author{
Turner Silverthorne
}
