% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfun_svdpower.R
\name{costfun_svdpower}
\alias{costfun_svdpower}
\title{Helper function for continuous power optimization}
\usage{
costfun_svdpower(mt, freqs, Amp = 1, alpha = 0.05)
}
\description{
Helper function for continuous power optimization
}
