% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfuns.R
\name{costfun_svdpower_discrete}
\alias{costfun_svdpower_discrete}
\title{Cost function for discrete-time power optimisation}
\usage{
costfun_svdpower_discrete(xinds, tau, freqs, ...)
}
\arguments{
\item{xinds}{the indices of \code{tau} at which measurements should be taken}

\item{tau}{the underlying fine grid from which measurements should be selected}

\item{freqs}{frequencies at which cost function should be evaluated}
}
\description{
Similar to \link{costfun_svdpower} except the points are chosen from a fixed
underlying grid, rather than allowed to be arbitrary points in the unit interval.
}
\author{
Turner Silverthorne
}
