% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimateTPR.R
\name{estimateTPR}
\alias{estimateTPR}
\title{Estimate TPR for multi-frequency hypothesis testing}
\usage{
estimateTPR(alpha, Nmc, tvec, Amp = 1, freq = 1, p_osc = 0.1, mc_cores = 1)
}
\arguments{
\item{alpha}{the false positive rate}

\item{Nmc}{number of Monte-Carlo replicates}

\item{tvec}{vector of measurement times}

\item{Amp}{amplitude of signal}

\item{freq}{frequency of signal}

\item{p_osc}{what proportion of the simulated data contains oscillations}

\item{mc_cores}{number of cores to be used for parallelization}
}
\value{
a dataframe containing estimate of the TPR for each combination FDR method
and hyptest given in \code{FDR_list} and \code{hyptest_list}
}
\description{
Estimate TPR for multi-frequency hypothesis testing
}
\author{
Turner Silverthorne
}
