% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/power_utils.R
\name{eval_montecarlo_power}
\alias{eval_montecarlo_power}
\title{Use Monte Carlo method to estimate power}
\usage{
eval_montecarlo_power(tvec, param, Nmc, alpha = 0.05)
}
\arguments{
\item{tvec}{vector of measurement times}

\item{param}{oscillation parameters
\itemize{
\item \code{Amp} amplitude of oscillation,
\item \code{freq} frequency of oscillation,
\item \code{acro} acrophase of oscillation.
}}

\item{Nmc}{number of Monte Carlo samples}

\item{alpha}{type I error, default value \code{alpha=.05}}
}
\value{
Monte Carlo estimate of power
}
\description{
Performs harmonic regression on a simulated dataset of independent samples
and returns the portion of samples that have statistically significant p-values.
Useful for comparison with the exact expression for statistical power,
see \code{\link[=eval_exact_power]{eval_exact_power()}}.
}
\author{
Turner Silverthorne
}
