% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalExactPower.R
\name{evalExactPower}
\alias{evalExactPower}
\title{Evaluate exact power of harmonic regression hypothesis test.}
\usage{
evalExactPower(t, param, alpha = 0.05)
}
\arguments{
\item{t}{measurement schedule}

\item{alpha}{type I error, by default \code{alpha=.05}}

\item{param$Amp}{amplitude of signal}

\item{param$freq}{frequency of signal}

\item{param$acro}{phase of signal in radians}
}
\value{
statistical power
}
\description{
Returns exact power of harmonic regression hypothesis test.
}
\note{
Assumes the noise has mean zero and unit standard deviation.
In these units, amplitude can be interpreted as the signal to noise ratio.
}
\author{
Turner Silverthorne
}
