% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWorstNCP.R
\name{evalWorstNCP}
\alias{evalWorstNCP}
\title{Eval worst-case non-centrality parameter}
\usage{
evalWorstNCP(mt, freqs, Amp = 1, alpha = 0.05, method = "svd")
}
\arguments{
\item{mt}{vector of measurement times}

\item{freqs}{vector of frequencies at which cost function should be evaluated}

\item{Amp}{amplitude of signal, only relevant if \code{cfuntype=='power'}}

\item{alpha}{type I error cutoff, only relevant if \code{cfuntype=='power'}}
}
\value{
minimum value of non-centrality parameter for freq in \code{freqs}
}
\description{
Eval worst-case non-centrality parameter
}
\author{
Turner Silverthorne
}
