% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalWorstPower.R
\name{evalWorstPower}
\alias{evalWorstPower}
\title{Eval worst-case power}
\usage{
evalWorstPower(mt, freqs, Amp = 1, alpha = 0.05, ...)
}
\arguments{
\item{mt}{vector of measurement times}

\item{freqs}{vector of frequencies at which cost function should be evaluated}

\item{Amp}{amplitude of signal, only relevant if \code{cfuntype=='power'}}

\item{alpha}{type I error cutoff, only relevant if \code{cfuntype=='power'}}
}
\value{
minimum value of non-centrality parameter for freq in \code{freqs}
}
\description{
Eval worst-case power
}
\author{
Turner Silverthorne
}
