% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_dominant_freq.R
\name{extract_dominant_freq}
\alias{extract_dominant_freq}
\title{Get dominant frequency from regression output}
\usage{
extract_dominant_freq(pqvals, amps, acros)
}
\arguments{
\item{pqvals}{\link{annmatrix} of p-values or q-values generated by \code{\link[=matrix_1d_padjust]{matrix_1d_padjust()}}}

\item{amps}{\link{annmatrix} of amplitudes, same shape as previous, each row corresponds
to a distinct experiment, each column to a distinct frequency}

\item{acros}{\link{annmatrix} of acrophases in units of radians

all three inputs to this function can be generated by \code{\link[=freqsweep_regr]{freqsweep_regr()}}}
}
\value{
estimated frequency, amplitude, acrophase, and p-value/q-value for each row
}
\description{
Helper function for processing output of \code{\link[=freqsweep_regr]{freqsweep_regr()}}. For each row,
this function selects the frequency with lowest p-value or q-value and returns the
corresponding amplitude and acrophase estimates and the p-value or q-value
}
\author{
Turner Silverthorne
}
