% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/freqsweep_regr.R
\name{freqsweep_regr}
\alias{freqsweep_regr}
\title{Independently fit cosinor models at all frequencies in a given range}
\usage{
freqsweep_regr(mt, Xdat, fmin, fmax, Nfreq, return_type)
}
\arguments{
\item{mt}{vector of measurement times}

\item{Xdat}{simulated dataset, for instance generated by \code{\link[=make_simulated_data]{make_simulated_data()}}}

\item{fmin}{lowest frequency to use in harmonic regression}

\item{fmax}{highest frequency to use in harmonic regression}

\item{return_type}{controls output of program, do you want only most
significant result or all results?}
}
\value{
a list with \code{\link[annmatrix:annmatrix]{annmatrix::annmatrix()}} entries
\itemize{
\item \code{pvalues} pvalue obtained for harmonic regression at each frequency
regression.
\item \code{amps} amplitude estimates for each frequency
\item \code{acros} acrophase estimates for each frequency, in units of radians
}
The individual harmonic regressions are performed using the
\code{\link[=rowCosinor]{rowCosinor()}} function.
}
\description{
Perform independent harmonic regressions at all frequencies in a given range.
}
\author{
Turner Silverthorne
}
