% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_utils.R
\name{getDeigDfreq}
\alias{getDeigDfreq}
\title{Frequency derivative of smallest FIM eigenvalue}
\usage{
getDeigDfreq(mt, freq)
}
\arguments{
\item{mt}{vector of measurement times}

\item{freq}{frequency of signal}
}
\value{
derivative of minimum eigenvalue of reduced FIM with respect to frequency
}
\description{
computes derivative of minimum eigenvalue of reduced Fisher information matrix
with respect to the frequency parameter
}
\author{
Turner Silverthorne
}
