% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_utils.R
\name{getMinEig}
\alias{getMinEig}
\title{Helper function returns smallest eigenvalue}
\usage{
getMinEig(Mat, is_symmetric = F)
}
\arguments{
\item{Mat}{a matrix}

\item{is_symmetric}{boolean, is the matrix symmetric}
}
\value{
smallest eigenvalue of mat, computed using \link[base:eigen]{base::eigen}
}
\description{
Wrapper for computing smallest eigenvalue of a matrix
}
\author{
Turner Silverthorne
}
