% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_exact_power.R
\name{getPower}
\alias{getPower}
\title{Evaluate exact power of harmonic regression hypothesis test.}
\usage{
getPower(t, param, alpha = 0.05)
}
\arguments{
\item{t}{measurement schedule}

\item{param}{list of oscillation parameters:
\itemize{
\item \code{Amp}  amplitude,
\item \code{freq} frequency,
\item \code{acro} acrophase.
}}

\item{alpha=.05}{type I error}
}
\value{
statistical power for given parameters
}
\description{
Returns exact power of harmonic regression hypothesis test.
}
\author{
Turner Silverthorne
}
