% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_auglattice_to_state}
\alias{helper_auglattice_to_state}
\title{Convert augmented lattice to a vector of measurement times}
\usage{
helper_auglattice_to_state(N1, N2, shift2, scale2)
}
\arguments{
\item{N1}{number of points in first lattice}

\item{N2}{number of points in second lattice}

\item{shift2}{offset of second lattice}

\item{scale2}{scale factor for second lattice}
}
\value{
vector of measurement times corresponding to the input 2-lattice parameterisation
}
\description{
Convert augmented lattice to a vector of measurement times
}
\author{
Turner Silverthorne
}
