% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_gap_penalty}
\alias{helper_gap_penalty}
\title{For penalizing cost function based on consecutive measurement gap}
\usage{
helper_gap_penalty(mt)
}
\arguments{
\item{mt}{vector of measurement times in unit interval \code{[0,1]}}
}
\value{
the largest gap between consecutive measurements.
}
\description{
For penalizing cost function based on consecutive measurement gap
}
\note{
This includes the gap between the final measurement and first measurement
\code{1-(mt[length(mt)]-mt[1])}
}
\author{
Turner Silverthorne
}
