% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_unif_update_scale_discrete}
\alias{helper_unif_update_scale_discrete}
\title{Discretised version of scale update}
\usage{
helper_unif_update_scale_discrete(scale, shift, Npts, tscale, Nfine)
}
\description{
Discretised version of scale update
}
\seealso{
\link{helper_unif_update_scale}
}
\author{
Turner Silverthorne
}
