% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_unif_update_shift}
\alias{helper_unif_update_shift}
\title{Helper function for lattice-based simulated annealing}
\usage{
helper_unif_update_shift(scale, shift, lat, tscale)
}
\arguments{
\item{scale}{current scale of lattice}

\item{shift}{current state of lattice}

\item{lat}{vector of measurement times in current lattice}

\item{tscale}{width parameter for uniform random points}
}
\value{
updates the shift of the lattice by sampling a uniform distribution
of width \code{tscale}, rejection sampling is then used to make sure that the
lattice with the updated shift parameter is still contained in the unit interval.
}
\description{
Helper function for lattice-based simulated annealing
}
\author{
Turner Silverthorne
}
