% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{helper_unif_update_shift_discrete}
\alias{helper_unif_update_shift_discrete}
\title{Discretised version of shift update}
\usage{
helper_unif_update_shift_discrete(scale, shift, Npts, tscale, Nfine)
}
\description{
Discretised version of shift update
}
\seealso{
\link{helper_unif_update_shift}
}
\author{
Turner Silverthorne
}
