% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsp_pseudowindow.R
\name{lsp_pseudowindow}
\alias{lsp_pseudowindow}
\title{Evaluate pseudo window of a Lomb Scargle periodogram}
\usage{
lsp_pseudowindow(mt, freqs, acros, type = "null")
}
\arguments{
\item{mt}{vector of measruement times}

\item{freqs}{frequencies at which to evaluate pseudo window}

\item{acros}{acrophases at which to evaluate pseudo-window}

\item{type}{a label for the output, useful for sweeping over multiple collections of measurement times}
}
\value{
a \link{data.frame} of Lomb-Scargle periodograms for each of the input
frequencies and acrophases
}
\description{
Given a frequency, acrophase, and collection of measurement times, the Lomb-Scargle
periodogram of this signal is referred to as the pseudo-window of the
measurement times

It is useful for estimating how accurately different signals will be detected
by the measurement times.
}
\author{
Turner Silverthorne
}
