% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_constraints.R
\name{make_constraints}
\alias{make_constraints}
\title{Generate CVXR constraints for power optimization}
\usage{
make_constraints(x, Lmat, bvec, opts)
}
\arguments{
\item{x}{the decision variable for power optimization generated by \code{\link[=make_variable]{make_variable()}}}

\item{Lmat}{matrix of lattice constraints, generated by \code{\link[=make_ineqmat]{make_ineqmat()}}}

\item{opts}{optimization options, could be generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
code(constraints) a list of constraints to be included when \code{\link[=make_problem]{make_problem()}} constructs the optimization problem.
}
\description{
\code{make_constraints()} creates a \code{\link[=list]{list()}} of constraints to be used by CVXR for power optimization.
}
\author{
Turner Silverthorne
}
