% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_default_opts.R
\name{make_default_opts}
\alias{make_default_opts}
\title{Generate default options for convex power optimization}
\usage{
make_default_opts(prob_size = "small")
}
\arguments{
\item{prob_size}{size of problem, either small, medium, or large}
}
\value{
\code{opts} a list containing the following optimization options:

related to optimization solver:
\itemize{
\item \code{opts$verbose} turn on/off intermediate output from Gurobi (\code{TRUE} by default)
\item \code{opts$lattice_cstr} decide if you want to constrain design space to only integer
combinations of certain sub-lattices  (\code{'none'} by default)
\item \code{opts$costfun_type} either \code{'L1'} which corresponds to averaging over frequencies
or \code{'Linfty'} which corresponds to maximizing power at worst case frequency
\item \code{opts$num_iter} iteration cutoff for GUROBI, value depends on \code{prob_size} input
}

related to time and frequency discretization
\itemize{
\item \code{opts$Nfine} number of candidate points to include when discretizing time
\item \code{opts$Nmeas} size of the measurement budget, acts as a constraint
\item \code{opts$Nfreq} number of frequencies to use in frequency discretization
\item \code{opts$fmin}  minimum frequency included in prior (for now only uniform prior allowed)
\item \code{opts$fmax} maximum frequency included in prior (for now only uniform prior allowed)
}

related to lattice construction (only relevant if \code{use_lattice==T}:
\itemize{
\item \code{opts$min_dx} smallest lattice spacing
\item \code{opts$max_lat_active} maximum number of lattices that can be used in constructing optimal solution
\item \code{opts$verbose maximum} number of lattices that can be used in constructing optimal solution
}
}
\description{
\code{make_default_opts()} generates optimization options for small medium or large problems.
When specifying to your own problem, some options should be kept
}
\author{
Turner Silverthorne
}
