% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ineqmat.R
\name{make_ineqmat}
\alias{make_ineqmat}
\title{Generate lattice constraints for convex power optimization}
\usage{
make_ineqmat(opts)
}
\arguments{
\item{opts}{optimization options generated by \code{\link[=make_default_opts]{make_default_opts()}}.}
}
\value{
\code{Lmat} a matrix of linear constraints on the sampling grid
\itemize{
\item if \code{opts$lattice_cstr=='none'} no lattice constraints are imposed (i.e. \code{Lmat=NULL})
\item if \code{opts$lattice_cstr=='lineq'}, each row of \code{Lmat} represents a constraint. The constraints will be added to the list of constraints generated by \code{\link[=make_constraints]{make_constraints()}}
\item if \code{opts$lattice_cstr=='cfun'}, each column of \code{Lmat} represents a constraint. The constraints will be incorporated in the cost function generated by \code{\link[=make_problem]{make_problem()}}
}

The third option tends to be considerably slower than the second even though they solve equivalent problems. It is
incldued here only for benchmarking and reference.
}
\description{
This function generates a matrix of linear constraints that ensure the optimal design can be
decomposed into at most \code{opts$max_lat_active} number of uniform grids.

Imposing these constraints may be useful if one seeks a design that balances performance with convenience of implementation.
}
\author{
Turner Silverthorne
}
