% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ineqrhs.R
\name{make_ineqrhs}
\alias{make_ineqrhs}
\title{Generate vector used in imposing lattice constraints}
\usage{
make_ineqrhs(Lmat, opts)
}
\arguments{
\item{Lmat}{a matrix of lattice constraints generated by \code{\link[=make_ineqmat]{make_ineqmat()}}}

\item{opts}{a list of optimization options, for instance generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
\code{bvec} a vector used in formulating the lattice constraints
\itemize{
\item if \code{opts$lattice_cstr=='cfun'} then \code{bvec} is the \code{colSums} of \code{Lmat}
\item if \code{opts$lattice_cstr=='cfun'} then \code{bvec} is the \code{rowSums} of \code{Lmat}
}
}
\description{
Given the constraint matrix and optimization options, \code{make_ineqvec()}
generates a vector used in formulating the lattice constraints.
}
\author{
Turner Silverthorne
}
