% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_quadmats.R
\name{make_quadmats}
\alias{make_quadmats}
\title{Generate matrices for quadratic forms used in convex power optimization}
\usage{
make_quadmats(Lmat, opts)
}
\arguments{
\item{Lmat}{lattice inequality matrix generated by \code{\link[=make_ineqmat]{make_ineqmat()}}}

\item{opts}{optimization options, defaults can be generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
\code{Aquad} either a single SPD matrix or a list of SPD matrices, with each matrix
corresponding to a different frequency under consideration.
}
\description{
The output of \code{make_quadmats()} depends on the choice of cost function specified in
\code{opts$costfun_type}.
\itemize{
\item If using an L1 cost function, the output will be an average of
several SPD matrices, with each matrix corresponding to a candidate frequency of interest.
\item If using an Linfty cost function, a list of SPD matrices will be returned suitable for a
downstream maximin convex programming.
}
}
\author{
Turner Silverthorne
}
