% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_utils.R
\name{make_simulated_data}
\alias{make_simulated_data}
\title{(Depracated) Generate simulated dataset with amplitude and freq uncertainty}
\usage{
make_simulated_data(mt, Nmc, Amin, Amax, fmin, fmax)
}
\arguments{
\item{mt}{vector of measurement times}

\item{Nmc}{size of simulated dataset}

\item{Amin}{minimum amplitude}

\item{Amax}{maximum amplitude}

\item{fmin}{minimum frequency}

\item{fmax}{maximum frequency}
}
\value{
an \code{\link[annmatrix:annmatrix]{annmatrix::annmatrix()}} object with rows labeled by amplitude,
frequency, and acrophase. Columns are labeled by measurement time \code{mt}.
}
\description{
Generated simulated dataset with uniformly random frequency and amplitude and
Gaussian white noise. The amplitude is assumed to be in units such that the noise
is of strength 1. In other words, the amplitude is the signal to noise ratio.
}
\author{
Turner Silverthorne
}
