% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvxr_utils.R
\name{make_variable}
\alias{make_variable}
\title{Generate decision variable for convex power optimization}
\usage{
make_variable(opts)
}
\arguments{
\item{opts$Nfine}{the number of points in the underlying fine grid from which
measurement times are to be chosen.}
}
\value{
\code{x} binary \link[CVXR:Variable-class]{CVXR::Variable} to be used as the decision variable in CVXR programming.
}
\description{
generates a binary vector which represents the weight function used by CVXR
for power optimization.
}
\author{
Turner Silverthorne
}
