% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_1d_padjust.R
\name{matrix_1d_padjust}
\alias{matrix_1d_padjust}
\title{Perform multiple test correction on rows or columns of a p-value matrix}
\usage{
matrix_1d_padjust(pdat, dim, pmethod)
}
\arguments{
\item{pdat}{matrix of pvalues}

\item{dim}{dimension along which adjustment should be performed
\itemize{
\item if \code{dim==1}, each row is treated as a collection of hypothesis tests for which
a set of q-values should be computed
\item if \code{dim==2}, each column is treated as a collection of hypothesis tests for which
a set of q-values should be computed
}}

\item{pmethod}{adjustment method, passed to \code{\link[stats:p.adjust]{stats::p.adjust()}}}
}
\value{
matrix of q-values
}
\description{
User chooses if multiple test correction should be performed row-wise or column-wise.
Depending on this choice, rows or columns are treated as the p-values from
independent hypothesis tests and are corrected using the \code{\link[stats:p.adjust]{stats::p.adjust()}} function.
}
\author{
Turner Silverthorne
}
