% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/opt_osc_power.R
\name{opt_osc_power}
\alias{opt_osc_power}
\title{Maximize worst-case power for cosinor-based rhythm detection}
\usage{
opt_osc_power(
  dvar0 = NULL,
  freqs,
  control,
  nlattice_opts = NULL,
  tau = NULL,
  ...
)
}
\description{
Template function for wrapping all optimization methods in suite.

svdpower with L-BFGS-B requires:
control$trace
control$REPORT
control$maxit
svdpower with simul_anneal requires:
same as previous, in additionlpha
control options necessary for specifying transition function
}
\author{
Turner Silverthorne
TODO add warning for passing args to control that should
appear explicitly in this function (i.e. cfuntype)
}
