% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_ineqmat.R
\name{prune_ineqmat}
\alias{prune_ineqmat}
\title{Remove redundant rows from lattice constraint matrix}
\usage{
prune_ineqmat()
}
\arguments{
\item{\code{Lmat}}{the inequality matrix, rows correspond to distinct constraints}
}
\value{
\code{Pmat} the pruned matrix
}
\description{
The constraint matrix generated by \code{\link[=make_ineqmat]{make_ineqmat()}} generally contains a large
number of redundant entries. This function removes redundancies by pruning away
rows of the matrix that are contained in the span of the other rows
}
\details{
Lattice constraints allow convex programming to restrict design space to
designs that can be easily decomposed into a family of uniform grids. If this
is implemented naively, the design space tends to be heavily over-parameterized,
leading to a very large matrix of linear lattice constraints.

To avoid this, we remove redundant rows from the matrix
}
\author{
Turner Silverthorne
}
