% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rowCosinor.R
\name{rowCosinor}
\alias{rowCosinor}
\title{Cosinor model}
\usage{
rowCosinor(theData, zts, per = 24)
}
\arguments{
\item{theData}{input matrix}

\item{zts}{time values for each column of theData}

\item{per}{period of oscillation to be fitted (default is 24)}
}
\value{
a data.frame containing various results of the fit including
p-values, R-squared values, acrophases and amplitudes for each row.
}
\description{
Function to fit a cosinor model for each row of the matrix.
}
\details{
Fits a cosinor model without any additional vocatiates.
}
\author{
Karolis Koncevicius
}
