% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_sa_power.R
\name{run_sa_power}
\alias{run_sa_power}
\title{Simulated annealing routine for lattice constraints in power maximization}
\usage{
run_sa_power(Aquad, opts)
}
\arguments{
\item{Aquad}{the matrix}

\item{opts}{options}
}
\value{
simulated annealing result
}
\description{
Use simulated annealing to generate candidate solutions to power maximization problem
}
\author{
Turner Silverthorne
}
