% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_cfunpwr.R
\name{sa_cfunpwr}
\alias{sa_cfunpwr}
\title{Cost function for simulated annealing power}
\usage{
sa_cfunpwr(x, Amat, opts)
}
\arguments{
\item{x}{binary lattice}

\item{Amat}{symmetric matrix/matrices defining optimization problem,
see \code{\link[=make_quadmats]{make_quadmats()}}}

\item{opts}{optimization options, generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
value of cost function
}
\description{
Evaluates cost function for simulated annealing power optimization.
See \code{\link[=run_sa_power]{run_sa_power()}} for details.
}
\author{
Turner Silverthorne
}
