% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_enumpar.R
\name{sa_enumpar}
\alias{sa_enumpar}
\title{Enumerate partitions of n}
\usage{
sa_enumpar(N)
}
\description{
Use pentagonal number recurrence relation to enumerate partitions of all integers in a range 1 to n
}
\author{
Turner Silverthorne
}
