% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_eulermat.R
\name{sa_eulermat}
\alias{sa_eulermat}
\title{Generate the probability matrices used for generating random partitions}
\usage{
sa_eulermat(m)
}
\description{
Helper function for \code{\link[=sa_randpar]{sa_randpar()}}, generates the probability matrices used in efficiently
sampling random partitions of a given integer using a combinatorial identity of Euler.
}
\author{
Turner Silverthorne
}
