% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_lat2state.R
\name{sa_lat2state}
\alias{sa_lat2state}
\title{Convert lattice structure to binary vector representing measurements}
\usage{
sa_lat2state(x, opts)
}
\arguments{
\item{x}{current state of system}

\item{opts}{optimization options, generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\description{
When using lattice constraints for simulated annealing, the state vector is
represented by a list of lattices. This function takes that list and constructs
a binary vector of length \code{opts$Nfine} representing the actual state of
the system.

This function is mostly used as a helper function for evaluating the simulated
annealing cost function found in \code{\link[=sa_cfunpwr]{sa_cfunpwr()}}.
}
\author{
Turner Silverthorne
}
