% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_propfunction.R
\name{sa_propfunction}
\alias{sa_propfunction}
\title{State proposal function for simulated annealing power maximization}
\usage{
sa_propfunction(x, opts)
}
\arguments{
\item{x}{state of system on current iteration of simulated annealing}

\item{opts}{simulation options, can be generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
\code{y} the new state of the system. Procedure for generating \code{y}
from input \code{x} depends on whether or not lattice constraints are imposed.
\itemize{
\item if \code{opts$lattice_cstr='none'} then \code{y} is obtained from \code{x} by randomly
swapping an active entry of \code{x} with an inactive entry of \code{x}
\item if \code{opts$lattice_cstr='sa_lattice'}, then \code{x} should be interpreted as a
superposition of lattices. A subset of the active lattices L' in L is selected at random to be
turned off. If there is a total of \code{n} points in this lattice, a partition of \code{n}
is chosen at random and a new set of uniform lattices consistent with this partition replaces
the original L' lattices.
}
}
\description{
Randomly generate new state for power maximization using simulated annealing.
You will most likely not need to use this function directly, it is called
by the function \code{\link[=run_sa_power]{run_sa_power()}}
}
\author{
Turner Silverthorne
}
