% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_randlattice.R
\name{sa_randlattice}
\alias{sa_randlattice}
\title{Random lattice for simulated annealing}
\usage{
sa_randlattice(n, opts)
}
\arguments{
\item{n}{number of measurements}

\item{opts$Nfine}{coarseness of binary lattice see \code{\link[=make_default_opts]{make_default_opts()}}}
}
\description{
Generates a random binary lattice for the simulated annealing transition function.
See \code{\link[=run_sa_power]{run_sa_power()}} for details. When there are lattice constraints, a more
complicated lattice proposal function is called in \code{\link[=sa_propfunction]{sa_propfunction()}}.
}
\author{
Turner Silverthorne
}
