% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_randpar.R
\name{sa_randpar}
\alias{sa_randpar}
\title{Generate uniformly random partition of integer n}
\usage{
sa_randpar(n, opts = NULL, parent_size = NULL)
}
\arguments{
\item{n}{non-negative integer which one wishes to partition}

\item{opts}{options to specify restrictions on partition, will be enforced using rejection sampling}
}
\value{
a uniformly random partition of \code{n}.
}
\description{
This function returns a uniformly random partition of the input integer \code{n}. The algorithm here
is from Combinatorial Algorithms by Nijenhuis and Wilf (1975).

It is a useful for constructing the transition function for running simulated annealing
with lattice constraints.
}
\author{
Turner Silverthorne
}
