% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_sample_inds.R
\name{sa_sample_inds}
\alias{sa_sample_inds}
\title{Helper for simulated annealing lattice proposal function}
\usage{
sa_sample_inds(m, opts)
}
\arguments{
\item{\code{m}}{total number of lattices in current state, see \code{\link[=sa_propfunction]{sa_propfunction()}}}

\item{\code{opts$lprop_method}}{determines what weighting will be used for
sampling a random subset of the lattices. The random subset is generated by
first choosing a subset size \code{k<=m} and the sampling \code{k} indices at random.
The subset size is determined randomly according to one of the following
distributions:
\itemize{
\item \code{opts$lprop_method='uniform'} choose \code{k~Unif(1,m)},
\item \code{opts$lprop_method='linear'} choose \code{k} with the weighting \code{[m,m-1,...,1]}, meaning that
smaller values of \code{k} are more likely,
\item \code{opts$lprop_method='exp'} choose \code{k} with the weighting \code{[2^{-k}]}
}
and then subsequent index sampling is always the same, random sampling without
replacement.}
}
\value{
a list of indices which represent the randomly sampled lattices.
}
\description{
Given a list of \code{m} lattices, this function returns a random subset of
the lattices.
}
\author{
Turner Silverthorne
}
