% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_test_for_overlap.R
\name{sa_test_for_overlap}
\alias{sa_test_for_overlap}
\title{Helper function for sa_propfunction}
\usage{
sa_test_for_overlap(xnew, x, inds, opts)
}
\description{
Helper function for sa_propfunction
}
\author{
Turner Silverthorne
}
