% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLombScarglePvals.R
\name{simLombScarglePvals}
\alias{simLombScarglePvals}
\title{Get p-values from Lomb Scargle periodogram}
\usage{
simLombScarglePvals(
  alpha_vals,
  Nmc,
  tvec,
  Amp = 1,
  freq = 1,
  p_osc = 0.1,
  mc_cores = 1
)
}
\arguments{
\item{alpha_vals}{values of threshold parameter}

\item{Nmc}{number of signals to include in dataset}

\item{tvec}{vector of measurement times}

\item{Amp}{amplitude of signal}

\item{freq}{frequency of signal}

\item{p_osc}{what proportion of the simulated data contains oscillations}

\item{mc_cores}{number of cores to be used for parallelization}
}
\value{
state vector indicating which genes are osc and which are not osc
a dataframe of simulated p-values
}
\description{
Get p-values from Lomb Scargle periodogram
}
\author{
Turner Silverthorne
}
