% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_cvxr.R
\name{solve_cvxr}
\alias{solve_cvxr}
\title{Solve power optimization using CVXR}
\usage{
solve_cvxr(control, Threads, ...)
}
\arguments{
\item{Threads}{how many threads should CVXR use}

\item{control$Nmeas}{number of measurements to be collected}

\item{control$Nfine}{number of points in fine grid from which measurement times are to be chosen}

\item{control$Nfreq}{number of points to use in frequency discretisation of objective}

\item{control$fmin}{minimum candidate frequency}

\item{control$fmax}{maximum candidate frequency}

\item{control$PreSolve}{recommended \code{PreSolve=2}, see \code{\link[CVXR:psolve]{CVXR::solve()}} for details}

\item{control$MIPFocus}{recommended \code{PreSolve=3}, see \code{\link[CVXR:psolve]{CVXR::solve()}} for details}

\item{control$cvxr_verbose}{boolean for how much output is desired from CVXR}

\item{control$num_iter}{max number of CVXR iterations}

\item{control$MIPGapAbs}{stopping criterion for CVXR solver}

\item{control$time_limit}{max time allowed for computation}

\item{control$prob_formualtion}{either max_elemwise, max_entries, or hvolume, see description}
}
\value{
the result of calling \link[CVXR:psolve]{CVXR::solve} with specified user controls
}
\description{
Performs DCP if \code{control$prob_formulation = 'max_elemwise'} or \code{'max_entries'}.
Performs hypervolume scalarization if \code{control$prob_formulation == 'hvolume'}.
}
\note{
the \code{...} argument is only relevant for computing the cost function value
at the end of this method, it is not relevant to the optimization stage.

\code{time_limit} does not count the time taken to complete the pre-solve.

if iterations exceed \code{num_iter}, it is difficultto access output of the computation.
For this reason, we suggest setting \code{num_iter} to an unreasonably high
value and using \code{time_limit} instead to determine the cut-off time for computation.
}
\seealso{
\code{\link[=opt_osc_power]{opt_osc_power()}} which contains a similar wrapper for CVXR
}
\author{
Turner Silverthorne
}
