% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/solve_pin2lat.R
\name{solve_pin2lat}
\alias{solve_pin2lat}
\title{Solve power optimization using simulated annealing and a pinned 2-lattice}
\usage{
solve_pin2lat(x0, freqs, control, ...)
}
\arguments{
\item{freqs}{list of frequencies for evaluating ojective function}

\item{x0$N1}{number of points in the roots of unity grid}

\item{x0$N2}{number of points in secondary grid}

\item{x0$shift2}{shift of secondary grid relative to start of experiment}

\item{x0$scale2}{scale of secondary grid}

\item{control$N1min}{min number of points allowed in N1 lattice, see \link{tfun_auglattice}}

\item{control$N1max}{max number of points allowed in N1 lattice, see \link{tfun_auglattice}}

\item{control$maxit}{how many iterations of simulated annealing to run.}

\item{control$trace}{determines how verbose simulated annealing should be}

\item{control$REPORT}{determines how verbose simulated annealing should be}

\item{cfuntype}{set to \code{'ncp'} if non-centrality parameter should be used
or \code{'power'} if power should be used in defining cost function.}
}
\value{
result of running simulated annealing using \link[stats:optim]{stats::optim}
}
\description{
Solve power optimization using simulated annealing and a pinned 2-lattice
}
\note{
additional inputs \code{...} are passed to \link{costfun_auglattice}
}
\seealso{
\link[stats:optim]{stats::optim} in the \code{method='SANN'} section for more information
on how iterations are counted. There may be multiple function calls within each iteration.
}
\author{
Turner Silverthorne
}
