% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfuns.R
\name{tfun_2lattice_svdpower}
\alias{tfun_2lattice_svdpower}
\title{Requires control$tfun_choice
control$mean_scale1
control$sd_scale1
control$mean_scale2
control$sd_scale2
control$mean_shift2
control$sd_shift2}
\usage{
tfun_2lattice_svdpower(shift1, shift2, scale1, scale2, lat1, lat2, control)
}
\description{
Requires control$tfun_choice
control$mean_scale1
control$sd_scale1
control$mean_scale2
control$sd_scale2
control$mean_shift2
control$sd_shift2
}
