% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfuns.R
\name{tfun_2lattice_svdpower_discrete}
\alias{tfun_2lattice_svdpower_discrete}
\title{requires control$tfun_choice}
\usage{
tfun_2lattice_svdpower_discrete(N1, dx1, N2, dx2, xshift2, Nfine, control)
}
\description{
requires control$tfun_choice
}
