% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tfuns.R
\name{tfun_svdpower}
\alias{tfun_svdpower}
\title{requires control$tfun_choice
control$tfun_mean
control$tfun_sd}
\usage{
tfun_svdpower(mt, control)
}
\description{
requires control$tfun_choice
control$tfun_mean
control$tfun_sd
}
