% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap_optimsa.R
\name{wrap_optimsa}
\alias{wrap_optimsa}
\title{Wrapper for optimising power directly using simulated annealing}
\usage{
wrap_optimsa(Nmeas, fmin, fmax, Nf, Nacro, Amp = 1)
}
\arguments{
\item{Nmeas}{measurement budget of experiment}

\item{fmin}{lowest frequency of interest}

\item{fmax}{highest frequency of interest}

\item{Nf}{number of points to use in freq discretisation}

\item{Nacro}{number of points to use in acrophase discretisation}

\item{Amp}{amplitude of signal}
}
\value{
output of simulated annealing
}
\description{
Wrapper for maximizing worst-case power using \code{\link[optimization:optim_sa]{optimization::optim_sa()}}. In
this function, measurement times are not restricted to a lattice. Worst-case
power is interpreted as the minimum power for any frequency between
\code{fmin} and \code{fmax}, and any acrophase in \code{[0,2*pi]}.
}
\author{
Turner Silverthorne
}
