% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{wrapper_sweep_lattice}
\alias{wrapper_sweep_lattice}
\title{Sweep over different partiions of a measurement budget and return optimal}
\usage{
wrapper_sweep_lattice(optim_routine, Nvals, Nmeas, cts_flag, dvar0, ...)
}
\arguments{
\item{optim_routine}{the optimization routine that should be run}

\item{Nvals}{which values of N1 should be considered}

\item{Nmeas}{total measurement budget}

\item{cts_flag}{is your optimization routine using continuous time?}

\item{dvar0}{initial state of optimizer}
}
\value{
a list containing the optimal solution, total computation time, and the
values of the objective function at all other Nvals for comparison.
}
\description{
Given an \code{Nmeas} budget, split \code{Nmeas=N1+N2} for various values of
\code{N1} and \code{N2} and perform optimization in this subspaces
}
\author{
Turner Silverthorne
}
