% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{xinds_from_lat1lat2_pars}
\alias{xinds_from_lat1lat2_pars}
\title{Helper function for discrete two-lattice optimization}
\usage{
xinds_from_lat1lat2_pars(N1, dx1, N2, dx2, xshift2)
}
\arguments{
\item{N1}{number of points in first lattice}

\item{dx1}{spacing of first lattice}

\item{N2}{number of points in second lattice}

\item{dx2}{spacing of second lattice}

\item{xshift2}{offset of second lattice}
}
\value{
a vector of indices corresponding to measurement times
}
\description{
Given a fine-uniform grid and a parameterization of a 2-lattice contained
in this grid, this function returns the grid indicest that correspond to this parameterization.
}
\note{
all parameters should be integers, as they are parameterising indicies
}
\author{
Turner Silverthorne
}
