% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_2lattice_to_state.R
\name{convert_2lattice_to_state}
\alias{convert_2lattice_to_state}
\title{Helper function converts parameterization of two lattices into single lattice}
\usage{
convert_2lattice_to_state(shift1, shift2, scale1, scale2, lat1, lat2)
}
\description{
Helper function converts parameterization of two lattices into single lattice
}
