% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfuns.R
\name{costfun_2lattice_svdpower}
\alias{costfun_2lattice_svdpower}
\title{Cost function for 2-lattice continuous-time power optimisation}
\usage{
costfun_2lattice_svdpower(
  shift1,
  shift2,
  scale1,
  scale2,
  lat1,
  lat2,
  freqs,
  ...
)
}
\arguments{
\item{shift1}{shift of first grid, must be between 0 and 1}

\item{shift2}{shift of second grid, must be between 0 and 1}

\item{scale1}{scale factor for first grid}

\item{scale2}{scale factor for second grid}

\item{lat1}{first uniform grid}

\item{lat2}{second uniform grid}

\item{freqs}{frequencies at which cost function should be evaluated}
}
\description{
Similar to \link{costfun_2lattice_svdpower_discrete} except time is no longer restricted
to a discrete grid; the lattices can be scaled arbitrarily.
}
\author{
Turner Silverthorne
}
