% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/costfuns.R
\name{costfun_svdpower}
\alias{costfun_svdpower}
\title{Cost function for continuous-time power optimisation}
\usage{
costfun_svdpower(
  mt,
  freqs,
  Amp = 1,
  alpha = 0.05,
  weight_ncp = 1,
  regL1 = 0,
  regFder = 0,
  gapPenalty = 0,
  leveragePenalty = 0,
  cfuntype = "power"
)
}
\arguments{
\item{mt}{vector of measurement times}

\item{freqs}{vector of frequencies at which cost function should be evaluated}

\item{Amp}{amplitude of signal, only relevant if \code{cfuntype=='power'}}

\item{alpha}{type I error cutoff, only relevant if \code{cfuntype=='power'}}

\item{weight_ncp}{factor for scaling non-centrality parameter in multi-objective scalarization.}

\item{regL1}{factor for scaling L1-frequency regularisation in multi-objective scalarization.}

\item{regFder}{factor for scaling frequency derivative in multi-objective scalarization.}

\item{gapPenalty}{factor for scaling adding a penalty for gaps between consecutive measurement times}

\item{leveragePenalty}{currently depracated, a factor for penalising by statistical leverage}

\item{cfuntype}{either 'power' or 'ncp' if non-centrality parameter is to be optimised directly}
}
\value{
minimum of cost function evaluated at \code{freqs}
}
\description{
Cost function for continuous-time power optimisation
}
\author{
Turner Silverthorne
}
