% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FIM_utils.R
\name{getReducedFIM}
\alias{getReducedFIM}
\title{Helper function computes reduced Fisher information matrix for cosinor model}
\usage{
getReducedFIM(t, param)
}
\description{
Here, reduced FIM correpsonds to de-meaned model.
}
