% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_problem.R
\name{make_problem}
\alias{make_problem}
\title{Generate a convex programming problem equivalent to power optimization}
\usage{
make_problem(x, Aquad, csts, opts)
}
\arguments{
\item{x}{a \code{\link[CVXR:Variable-class]{CVXR::Variable()}} representing the weight function for measurement times, see \code{\link[=make_variable]{make_variable()}}}

\item{Aquad}{the symmetric matrix/matrices that define the quadratic form to be optimized, see \code{\link[=make_quadmats]{make_quadmats()}}}

\item{csts}{a list made up of integer and convex constraints on \code{x}, see \code{\link[=make_constraints]{make_constraints()}}}

\item{opts}{a complete set of optimization options, default values can be generated by \code{\link[=make_default_opts]{make_default_opts()}}}
}
\value{
\code{prob} a disciplined convex programming problem which can be solved using the \link{CVXR} library.
\itemize{
\item See \code{\link[CVXR:Problem-class]{CVXR::Problem()}} for more details on this datatype.
}
}
\description{
\code{make_problem()} generates a disciplined convex programming problem that is
equivalent to maximizing the statistical power of a cosinor-based hypothesis test.
The problem is created as a \code{\link[CVXR:Problem-class]{CVXR::Problem()}} which can then be solved using \link{CVXR}
together with the \link{GUROBI} backend. The \link{GUROBI} backend is necessary because there
are integer constraints on the problem.
}
\author{
Turner Silverthorne
}
