% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvxr_utils.R
\name{make_quadmats}
\alias{make_quadmats}
\title{Generate matrices for quadratic forms used in convex power optimization}
\usage{
make_quadmats(opts, ...)
}
\arguments{
\item{opts$Nfine}{number of points in the fine time grid}

\item{opts$Nfreq}{number of frequencies in frequency grid}

\item{opts$Nmeas}{measurement budget}

\item{opts$fmin}{minimum frequency to be included}

\item{opts$fmax}{maximum frequency to be included}
}
\value{
a list of matrices of length \code{opts$Nfreq}. Each entry in the list
corresponds to a quadratic form at a frequency between \code{opts$fmin} and
\code{opts$fmax}.
}
\description{
Generate matrices for quadratic forms used in convex power optimization
}
\author{
Turner Silverthorne
}
