% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvxr_utils.R
\name{make_quadmats}
\alias{make_quadmats}
\title{Generate matrices for quadratic forms used in convex power optimization}
\usage{
make_quadmats(opts, returnFder = F, ...)
}
\arguments{
\item{opts}{optimization options, defaults can be generated by \code{\link[=make_default_opts]{make_default_opts()}}}

\item{Lmat}{NULL by default, lattice inequality matrix generated by \code{\link[=make_ineqmat]{make_ineqmat()}}}
}
\value{
\code{Aquad} a symmetric matrix to be used in constructing the quadratic form(s)
to be optimized in the convex programming routine.
\itemize{
\item If using an L1 cost function, the output will be an average of
several symmetric matrices, with each matrix corresponding to a candidate frequency of interest.
\item If using an Linfty cost function, a list of SPD matrices will be returned suitable for a
downstream maximin convex programming.
}
}
\description{
The output of \code{make_quadmats()} depends on the choice of cost function specified in
\code{opts$costfun_type}. If using an L1 cost function, different frequencies will be
averaged over in constructing the output matrix. If using an Linfty cost function, a
list of matrices will be returned, each matrix corresponding to a different frequency
under consideration.
}
\author{
Turner Silverthorne
}
