% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_variable.R
\name{make_variable}
\alias{make_variable}
\title{Generate decision variable for convex power optimization}
\usage{
make_variable(opts)
}
\arguments{
\item{opts}{optimization options generated by \code{\link[=make_default_opts]{make_default_opts()}}.
\itemize{
\item if \code{opts$lattice_cstr=='none'}, \code{x} will have \code{opts$Nfine} components
\item if \code{opts$lattice_cstr=='lineq'}, \code{x} will have \code{opts$Nfine} components
\item if \code{opts$lattice_cstr=='cfun'}, \code{x} will have as many dimensions as there are constraints in \code{Lmat}
}

In the first two cases, \code{x} represents as a weight function on the candidate measurement times. In the third case,
\code{x} represents as a weight function on the candidate sub-lattices.}

\item{Lmat}{a matrix containing the lattice constraints, only relevant to this function if \code{opts$lattice_cstr=='cfun'}}
}
\value{
\code{x} binary \link[CVXR:Variable-class]{CVXR::Variable} to be used as the decision variable in CVXR programming.
}
\description{
\code{make_variable()} generates a binary vector which represents the weight function used by CVXR
for power optimization.
}
\author{
Turner Silverthorne
}
