% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sa_cfunpwr.R
\name{sa_cfunpwr}
\alias{sa_cfunpwr}
\title{Cost function for simulated annealing power}
\usage{
sa_cfunpwr(x, Amat, opts)
}
\description{
Cost function for simulated annealing power
}
\author{
Turner Silverthorne
}
